backend_django/requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
Django==5.0.3
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
future==1.0.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openapi-codec==1.3.2
openpyxl==3.1.5
paramiko==3.5.1
presto-python-client==0.8.4
psycopg2-binary==2.9.10
py4j==0.10.9.7
pycparser==2.22
PyHive==0.7.0
PyJWT==2.9.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
sshtunnel==0.4.0
thrift==0.21.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
